+++ src/features/Onboarding/Common/index.tsx
@@ -8,8 +8,8 @@ import { Navigate } from 'react-router-dom';
 import Loading from '@/components/Loading/BrandTextLoading';
 import OnboardingContainer from '@/routes/onboarding/_layout';
 import { deriveOnboardingBranchPath } from '@/routes/onboarding/config';
-import IntroLanguageStep from '@/routes/onboarding/features/IntroLanguageStep';
-import PrivacyStep from '@/routes/onboarding/features/PrivacyStep';
+import ResponseLanguageStep from '@/routes/onboarding/features/ResponseLanguageStep';
+import TelemetryStep from '@/routes/onboarding/features/TelemetryStep';
 import { useServerConfigStore } from '@/store/serverConfig';
 import { useUserStore } from '@/store/user';
 import { onboardingSelectors } from '@/store/user/selectors';
@@ -22,18 +22,18 @@ const CommonOnboardingPage = memo(() => {
 
   const [step, setStep] = useState<1 | 2>(1);
 
-  const goNextFromIntro = useCallback(() => {
+  const goNextFromTelemetry = useCallback(() => {
     setStep(2);
   }, []);
 
-  const goBackFromPrivacy = useCallback(() => {
+  const goBackFromLanguage = useCallback(() => {
     setStep(1);
   }, []);
 
   const finishCommon = useCallback(() => {
-    // No-op: completion of step 2 writes telemetry, which flips
-    // commonStepsCompleted to true; the early-return below then
-    // handles the redirect on the next render.
+    // No-op: completion of step 2 writes responseLanguage, which flips
+    // commonStepsCompleted to true; the early-return below then handles
+    // the redirect on the next render.
   }, []);
 
   if (!isUserStateInit) {
@@ -55,9 +55,9 @@ const CommonOnboardingPage = memo(() => {
     <OnboardingContainer>
       <Flexbox gap={24} style={{ maxWidth: 600, width: '100%' }}>
         {step === 1 ? (
-          <IntroLanguageStep onNext={goNextFromIntro} />
+          <TelemetryStep onNext={goNextFromTelemetry} />
         ) : (
-          <PrivacyStep onBack={goBackFromPrivacy} onNext={finishCommon} />
+          <ResponseLanguageStep onBack={goBackFromLanguage} onNext={finishCommon} />
         )}
       </Flexbox>
     </OnboardingContainer>
